ISSUE-760 address review comments
